lark: do not export lark.shell_quote()
It is a tricky function to actually use an it is not relied upon other
than for reporting.  Any users needing such functionality would most
likely be better off writing their own version or just using a function
to quote single strings and table.concat() to construct the final
script.