gluatest: each test is isolated in a separate lua.LState
If global functions __test_setup and __test_teardown are defined they
are used to perform setup and teardown surrounding the execution of each
test function.

Any error encountered during a setup function is considered fatal and
the test function will not execute.  When a teardown function is defined
it will execute regardless of errors in the setup function.