repl: hide internal modules by encoding internal nature in import paths
Any package import path with a base name that begins with underscore '_'
is considered internal.  The package name itself is not expected to have
an underscore as tools will consider this a warning.

Nesting of packages under internal packages is not considered.  Such
packages should also have an underscore prefix to be considered
internal.